backend/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
bleach==4.1.0
certifi==2024.12.14
charset-normalizer==3.4.0
Django==5.1.4
django-cors-headers==4.6.0
django-jazzmin==3.0.1
django-summernote==0.8.20.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
packaging==24.2
pillow==11.0.0
PyJWT==2.10.1
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
urllib3==2.2.3
validators==0.34.0
webencodings==0.5.1
